chore: move tradeData state to formcomponent to prevent badstate
